# MeetupAPI/docs/api.yaml
openapi: 3.0.4
info:
  title: MeetupAPI
  version: v1
paths:
  /api/account/login:
    post:
      tags:
        - Account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
      responses:
        '200':
          description: OK
  /api/account/register:
    post:
      tags:
        - Account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
      responses:
        '200':
          description: OK
  /api/account/edit:
    put:
      tags:
        - Account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: OK
  /config/reload:
    options:
      tags:
        - Config
      responses:
        '200':
          description: OK
  /file:
    get:
      tags:
        - File
      parameters:
        - name: name
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
  '/api/meetup/{meetupName}/lecture':
    get:
      tags:
        - Lecture
      parameters:
        - name: meetupName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
    post:
      tags:
        - Lecture
      parameters:
        - name: meetupName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LectureRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/LectureRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/LectureRequest'
      responses:
        '200':
          description: OK
    delete:
      tags:
        - Lecture
      parameters:
        - name: meetupName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  '/api/meetup/{meetupName}/lecture/{id}':
    delete:
      tags:
        - Lecture
      parameters:
        - name: meetupName
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
  /api/meetup:
    get:
      tags:
        - Meetup
      parameters:
        - name: SearchPhrase
          in: query
          required: true
          schema:
            minLength: 1
            type: string
        - name: PageSize
          in: query
          schema:
            type: integer
            format: int32
        - name: PageNumber
          in: query
          schema:
            minimum: 1
            type: integer
            format: int32
        - name: SortBy
          in: query
          schema:
            type: string
            nullable: true
        - name: SortDirection
          in: query
          schema:
            $ref: '#/components/schemas/SortDirection'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MeetupResponsePagedResult'
            application/json:
              schema:
                $ref: '#/components/schemas/MeetupResponsePagedResult'
            text/json:
              schema:
                $ref: '#/components/schemas/MeetupResponsePagedResult'
    post:
      tags:
        - Meetup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeetupRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/MeetupRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/MeetupRequest'
      responses:
        '200':
          description: OK
  '/api/meetup/{name}':
    get:
      tags:
        - Meetup
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MeetupResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/MeetupResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/MeetupResponse'
    put:
      tags:
        - Meetup
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeetupRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/MeetupRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/MeetupRequest'
      responses:
        '200':
          description: OK
    delete:
      tags:
        - Meetup
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
components:
  schemas:
    INotification:
      type: object
      additionalProperties: false
    Lecture:
      type: object
      properties:
        id:
          type: integer
          format: int32
        author:
          type: string
          nullable: true
        topic:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        meetup:
          $ref: '#/components/schemas/Meetup'
        meetupId:
          type: string
          format: uuid
      additionalProperties: false
    LectureRequest:
      required:
        - author
        - topic
      type: object
      properties:
        author:
          minLength: 5
          type: string
        topic:
          minLength: 5
          type: string
        description:
          type: string
          nullable: true
      additionalProperties: false
    Location:
      type: object
      properties:
        id:
          type: integer
          format: int32
        city:
          type: string
          nullable: true
        street:
          type: string
          nullable: true
        postCode:
          type: string
          nullable: true
        meetup:
          $ref: '#/components/schemas/Meetup'
        meetupId:
          type: string
          format: uuid
      additionalProperties: false
    Meetup:
      type: object
      properties:
        domainEvents:
          type: array
          items:
            $ref: '#/components/schemas/INotification'
          nullable: true
          readOnly: true
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        organizer:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
        isPrivate:
          type: boolean
        location:
          $ref: '#/components/schemas/Location'
        lectures:
          type: array
          items:
            $ref: '#/components/schemas/Lecture'
          nullable: true
        createdById:
          type: integer
          format: int32
          nullable: true
        createdBy:
          $ref: '#/components/schemas/User'
      additionalProperties: false
    MeetupQuery:
      required:
        - searchPhrase
      type: object
      properties:
        searchPhrase:
          minLength: 1
          type: string
        pageSize:
          type: integer
          format: int32
        pageNumber:
          minimum: 1
          type: integer
          format: int32
        sortBy:
          type: string
          nullable: true
        sortDirection:
          $ref: '#/components/schemas/SortDirection'
      additionalProperties: false
    MeetupRequest:
      required:
        - name
        - organizer
      type: object
      properties:
        name:
          minLength: 3
          type: string
        organizer:
          minLength: 1
          type: string
        date:
          type: string
          format: date-time
        isPrivate:
          type: boolean
      additionalProperties: false
    MeetupResponse:
      type: object
      properties:
        fullLocation:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        organizer:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
        isPrivate:
          type: boolean
        lectures:
          type: array
          items:
            $ref: '#/components/schemas/Lecture'
          nullable: true
      additionalProperties: false
    MeetupResponsePagedResult:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MeetupResponse'
          nullable: true
        totalPages:
          type: integer
          format: int32
        itemsFrom:
          type: integer
          format: int32
        itemsTo:
          type: integer
          format: int32
        totalItemsCount:
          type: integer
          format: int32
      additionalProperties: false
    RegisterUserRequest:
      type: object
      properties:
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        confirmPassword:
          type: string
          nullable: true
        nationality:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          format: date-time
          nullable: true
        roleId:
          type: integer
          format: int32
      additionalProperties: false
    Role:
      type: object
      properties:
        id:
          type: integer
          format: int32
        roleName:
          type: string
          nullable: true
      additionalProperties: false
    SortDirection:
      enum:
        - 0
        - 1
      type: integer
      format: int32
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        confirmPassword:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        nationality:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          format: date-time
          nullable: true
        roleId:
          type: integer
          format: int32
      additionalProperties: false
    User:
      type: object
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        nationality:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          format: date-time
          nullable: true
        passwordHash:
          type: string
          nullable: true
        roleId:
          type: integer
          format: int32
        role:
          $ref: '#/components/schemas/Role'
      additionalProperties: false